only loaded cart if a valid customer id is set

+++ Documents/NetBeansProjects/IoTBay001/web/cart.jsp
@@ -30,52 +30,55 @@
     System.out.println("cid: " + customer_id);
     
     // check if a new order needs to be created
-    ResultSet res = st.executeQuery("select * from orders where order_id not in (select order_id from payment) and customer_id = " + customer_id);
-
-    if (res.next() && order_id != null) {
-        System.out.println("oid: " + res.getString("order_id"));
-        request.getSession().setAttribute("order_id", res.getString("order_id"));
-    } else {
-        System.out.println("need to create an oid");
-
-        SimpleDateFormat date = new SimpleDateFormat("yyyy-MM-dd");
-        String date_ordered = date.format(new Date());
-
-        st.executeUpdate("insert into orders (date_ordered, customer_id) values ('" + date_ordered + "', " + customer_id + ")");
-        
+    ResultSet res; 
+    
+    if (!customer_id.equals("null")) {
         res = st.executeQuery("select * from orders where order_id not in (select order_id from payment) and customer_id = " + customer_id);
-        res.next();
-        request.getSession().setAttribute("order_id", res.getString("order_id"));
-        order_id = request.getSession().getAttribute("order_id") + "";
-    }
 
-    // check if item needs to be added
-    String product_id = request.getParameter("p");
-    String quantity = request.getParameter("q");
+        if (res.next() && !order_id.equals("")) {
+            System.out.println("oid: " + res.getString("order_id"));
+            request.getSession().setAttribute("order_id", res.getString("order_id"));
+        } else {
+            System.out.println("need to create an oid");
 
-    if (!order_id.equals("")) {
-        if (product_id != null && quantity != null) {
-            System.out.println("item needs to be added, adding: " + order_id + ", " + product_id + ", " + quantity);
-            st.executeUpdate("insert into orderlineitem (order_id, product_id, quantity) values (" + order_id + ", " + product_id + ", " + quantity + ")");
-            st.executeUpdate("update product set total_quantity = total_quantity - " + quantity + " where product_id = " + product_id);
+            SimpleDateFormat date = new SimpleDateFormat("yyyy-MM-dd");
+            String date_ordered = date.format(new Date());
 
-            // refresh page
+            st.executeUpdate("insert into orders (date_ordered, customer_id) values ('" + date_ordered + "', " + customer_id + ")");
+
+            res = st.executeQuery("select * from orders where order_id not in (select order_id from payment) and customer_id = " + customer_id);
+            res.next();
+            request.getSession().setAttribute("order_id", res.getString("order_id"));
+            order_id = request.getSession().getAttribute("order_id") + "";
+        }
+
+        // check if item needs to be added
+        String product_id = request.getParameter("p");
+        String quantity = request.getParameter("q");
+
+        if (!order_id.equals("")) {
+            if (product_id != null && quantity != null) {
+                System.out.println("item needs to be added, adding: " + order_id + ", " + product_id + ", " + quantity);
+                st.executeUpdate("insert into orderlineitem (order_id, product_id, quantity) values (" + order_id + ", " + product_id + ", " + quantity + ")");
+                st.executeUpdate("update product set total_quantity = total_quantity - " + quantity + " where product_id = " + product_id);
+
+                // refresh page
+                %><script>location.replace("cart.jsp")</script><%
+            }
+
+            // check if item needs to be removed
+            String removeProduct = request.getParameter("rp");
+            String removeQuantity = request.getParameter("rq");
+
+            if (removeProduct != null && removeQuantity != null) {
+                System.out.println("item needs to be removed");
+                st.executeUpdate("delete from orderlineitem where order_id = " + order_id + " and product_id = " + removeProduct + " and quantity = " + removeQuantity);
+                st.executeUpdate("update product set total_quantity = total_quantity + " + removeQuantity + " where product_id = " + removeProduct);
+            }
+        } else {
             %><script>location.replace("cart.jsp")</script><%
         }
-
-        // check if item needs to be removed
-        String removeProduct = request.getParameter("rp");
-        String removeQuantity = request.getParameter("rq");
-
-        if (removeProduct != null && removeQuantity != null) {
-            System.out.println("item needs to be removed");
-            st.executeUpdate("delete from orderlineitem where order_id = " + order_id + " and product_id = " + removeProduct + " and quantity = " + removeQuantity);
-            st.executeUpdate("update product set total_quantity = total_quantity + " + removeQuantity + " where product_id = " + removeProduct);
-        }
-    } else {
-        %><script>location.replace("cart.jsp")</script><%
     }
-    
 %>
 
 <!DOCTYPE html>
@@ -117,7 +120,7 @@
             <h1 class='align-middle' style='text-align: center;'>Cart</h1>
             <table>
                 <%
-                    if (!order_id.equals("")) {
+                    if (!order_id.equals("null")) {
                         
                         res = st.executeQuery("select product_id, order_id, product_name, quantity, price_per_unit, sum(quantity) total "
                                             + "from shoppingcart where order_id = " + order_id + " "
@@ -157,15 +160,23 @@
                             <h5 style="text-align: center"><a href="customerProductList.jsp">Click here</a> to check our catalogue!</h5>
                         <%
                         }
+                    } else {
+                        %><h1 class="text-center">You are currently signed out, <a href="login.jsp">click here</a> to log in.</h1><%
                     }
                     %>
             </table>
+            <%
+                if (!order_id.equals("null")) {
+            %>
             <form action="createOrder.jsp" method="POST">
                 <div class="mt-2 mr-2" style="text-align: right">
                     <input type="submit" class="btn btn-default bg-primary text-white mr-2" value="Buy Now"/>
                     <a class="btn btn-default bg-light" style="color: #000" href="customerProductList.jsp">Continue Shopping</a>
                 </div>                        
             </form>
+            <%
+                }
+            %>
         </div>
     </body>
 </html>
